refactor: découpage pdfExport.ts en modules (phase 2)
pdfExport.ts (1940 l.) → barrel de 39 l. + 8 modules (core, poolPdf, tableauPdf, bracketTreePdf, rankingPdf, resultsPdf, appelPdf, fullCompetitionPdf). API publique strictement identique, 11 appelants inchangés, 409 tests verts.
